% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nucleoSim.R
\docType{package}
\name{nucleoSim-package}
\alias{nucleoSim}
\alias{nucleoSim-package}
\title{nucleoSim: Generate a synthetic nucleosome map}
\description{
This package generate a synthetic nucleosome map using the parameters
given by the user. The user has choice between three different
distributions: Normal, Student and Uniform.
}
\author{
Rawane Samb,
Astrid Deschenes,
Pascal Belleau and
Arnaud Droit

Maintainer:
Astrid Deschenes <astrid-louise.deschenes@crchudequebec.ulaval.ca>
}
\seealso{
\itemize{
    \item \code{\link{syntheticNucMapFromDist}} {to generate a synthetic
nucleosome map}
}
}
\keyword{package}

