% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nucleoSimInternFunctions.R
\name{syntheticNucMapFromDistValidation}
\alias{syntheticNucMapFromDistValidation}
\title{Parameter validation for the \code{\link{syntheticNucMapFromDist}}
function}
\usage{
syntheticNucMapFromDistValidation(wp.num, wp.del, wp.var, fuz.num, fuz.var,
  max.cover, nuc.len, lin.len, rnd.seed, as.ratio, show.plot)
}
\arguments{
\item{wp.num}{the number of well-positioned
(non overlapping) nucleosomes.}

\item{wp.del}{the number of well-positioned nucleosomes to remove to create
uncovered regions.}

\item{wp.var}{the maximum variance in basepairs of the well-positioned
nucleosomes. This parameter introduces some variation in the position of
the reads describing a nucleosome.}

\item{fuz.num}{the number of fuzzy nucleosomes. Those nucleosomes are
distributed accordingly to selected distribution all over the region.
Those nucleosomes can overlap other well-positioned or fuzzy nucleosomes.}

\item{fuz.var}{the maximum variance of the fuzzy nucleosomes. This variance
can be different than the one used for the well-positioned
nucleosome reads.}

\item{max.cover}{the maximum coverage for one nucleosome. The final coverage
can have a higher value than \code{max.cover} since reads from differents
nucleosomes can be overlapping.}

\item{nuc.len}{the nucleosome length.}

\item{lin.len}{the length of the DNA linker DNA.}

\item{rnd.seed}{a single value, interpreted as an \code{integer}, or
\code{NULL}. If a \code{integer} is given, the value is used to set the seed
of the random number generator. By fixing the seed, the generated results
can be reproduced.}

\item{as.ratio}{a \code{logical}, if \code{TRUE},  a synthetic naked DNA
control map is created and the ratio between it and the nucleosome coverage
are calculated. It can be used to simulate hybridization ratio data, like
the one in Tiling Arrays. Both control map and calculated ratio are
returned.}

\item{show.plot}{a \code{logical}, if \code{TRUE}, a plot of the coverage
map, with the nucleosome calls and optionally the calculated ratio is
outputed.}
}
\value{
\code{0} indicating that all parameters validations have been
successful.
}
\description{
Parameter validation for the \code{\link{syntheticNucMapFromDist}}
function
}
\author{
Astrid Louise Deschenes
}
\keyword{internal}

