% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleoSimInternFunctions.R
\name{syntheticNucReadsValidation}
\alias{syntheticNucReadsValidation}
\title{Subsection of parameter validation for identical parameters between
\code{syntheticNucReadsFromMap} and \code{syntheticNucReadsFromDist}
functions.}
\usage{
syntheticNucReadsValidation(read.len, offset)
}
\value{
\code{0} indicating that all parameters validations have been
successful.
}
\description{
Validate that identical values passed to both
\code{syntheticNucReadsFromMap} and \code{syntheticNucReadsFromDist}
functions are correctly formatted.
}
\author{
Astrid Deschenes
}
\keyword{internal}

