% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nucleoSimFunctions.R
\name{syntheticSampleFromDist}
\alias{syntheticSampleFromDist}
\title{Generate a synthetic nucleosome map containing forward and
reverse reads}
\usage{
syntheticSampleFromDist(wp.num, wp.del, wp.var, fuz.num, fuz.var,
  max.cover = 100, nuc.len = 147, len.var = 10, lin.len = 20,
  read.len = 40, rnd.seed = NULL, as.ratio = FALSE, distr = c("Uniform",
  "Normal", "Student"), offset)
}
\arguments{
\item{wp.num}{the number of well-positioned
(non overlapping) nucleosomes.}

\item{wp.del}{the number of well-positioned nucleosomes to remove to create
uncovered regions.}

\item{wp.var}{the maximum variance in basepairs of the well-positioned
nucleosomes. This parameter introduces some variation in the position of
the reads describing a nucleosome.}

\item{fuz.num}{the number of fuzzy nucleosomes. Those nucleosomes are
distributed accordingly to selected distribution all over the region.
Those nucleosomes can overlap other well-positioned or fuzzy nucleosomes.}

\item{fuz.var}{the maximum variance of the fuzzy nucleosomes. This variance
can be different than the one used for the well-positioned
nucleosome reads.}

\item{max.cover}{the maximum coverage for one nucleosome. The final coverage
can have a higher value than \code{max.cover} since reads from differents
nucleosomes can be overlapping. Default = 100.}

\item{nuc.len}{the nucleosome length. Default = 147.}

\item{len.var}{a \code{numeric}, the variance of the distance between a
forward read and its paired reverse read. Default = 10.}

\item{lin.len}{the length of the DNA linker DNA. Default = 20.}

\item{read.len}{the length of each of the paired reads. Default = 40.}

\item{rnd.seed}{a single value, interpreted as an \code{integer}, or
\code{NULL}. If a \code{integer} is given, the value is used to set the seed
of the random number generator. By fixing the seed, the generated results
can be reproduced. Default = \code{NULL}.}

\item{as.ratio}{a \code{logical}, if \code{TRUE},  a synthetic naked DNA
control map is created and the ratio between it and the nucleosome coverage
are calculated. It can be used to simulate hybridization ratio data, like
the one in Tiling Arrays. Both control map and calculated ratio are
returned. Default = \code{FALSE}.}

\item{distr}{the name of the distribution used to generate the nucleosome
map. The choices are : \code{"Uniform"}, \code{"Normal"} and
\code{"Student"}. Default = \code{"Uniform"}.}

\item{offset}{the number of bases used to offset all reads. This is done to
ensure that all nucleosome positions and read alignment are of positive
values.}
}
\value{
an \code{list} of \code{class} "nucleoSim" containing the
following elements:
\itemize{
\item \code{call} the matched call.
\item \code{dataIP} a \code{data.frame} with all forward and reverse reads
for all well-positioned and fuzzy nucleosomes.
\item \code{wp} a \code{data.frame} with the positions of all the
well-positioned nucleosomes, as well as the number of paired-reads
associated to each one.
\item \code{fuz} a \code{data.frame} with the positions of all the fuzzy
nucleosomes, as well as the number of paired-reads associated to each one.
\item \code{paired} a \code{data.frame} with the starting and ending
positions of the reads used to generate the paired-end reads.
}
}
\description{
Generate a synthetic nucleosome map using the distribution
selected by the user. The user has choice between three different
distributions: Normal, Student and Uniform.

This function is using a modified version of the syntheticNucMap() function
from Bioconductor nucleR package (Flores and Orozco, 2011).
}
\examples{
## Generate a synthetic map with 20 well-positioned + 10 fuzzy nucleosomes
## using a Normal distribution with a variance of 30 for the well-positioned
## nucleosomes, a variance of 40 for the fuzzy nucleosomes and a seed of 15
res <- syntheticSampleFromDist(wp.num = 20, wp.del = 0, wp.var = 30,
fuz.num = 10, fuz.var = 40, as.ratio = TRUE, rnd.seed = 15, distr = "Normal",
offset = 1000)
}
\author{
Pascal Belleau, Rawane Samb, Astrid Deschenes
}

