% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleoSim.R
\docType{package}
\name{nucleoSim-package}
\alias{nucleoSim-package}
\alias{nucleoSim}
\title{nucleoSim: Generate synthetic nucleosome maps}
\description{
This package can generate a synthetic map with
reads covering the nucleosome regions as well as a synthetic map with
forward and reverse reads emulating next-generation sequencing.
The user has choice between three different
distributions for the read positioning: Normal, Student and Uniform.
}
\seealso{
\itemize{
    \item \code{\link{syntheticNucMapFromDist}} {to generate a synthetic
nucleosome map}
}
}
\author{
Rawane Samb,
Astrid Deschenes,
Pascal Belleau and
Arnaud Droit

Maintainer:
Astrid Deschenes <adeschen@hotmail.com>
}
\keyword{package}
